requirements.txt
# Core Dependencies
streamlit>=1.28.0
python-dotenv>=0.19.0

# AI/ML Dependencies
ollama>=0.1.7
pydantic>=2.0.0

# Web/API Dependencies
requests>=2.28.0
youtube-transcript-api>=0.6.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Audio/Video Processing
yt-dlp>=2023.12.30

# Professional Transcription (Optional)
assemblyai>=0.20.0

# PDF Generation
reportlab>=3.6.0

# Data Processing
pandas>=1.5.0
numpy>=1.24.0

# Configuration
PyYAML>=6.0

# Development Dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0
